zplus-engine-develop/requirements.txt
aiosignal==1.3.1
aspose-words==23.9.0
asyncio==3.4.3
attrs==23.1.0
bcrypt==4.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.4
defusedxml==0.7.1
dnspython==2.4.2
docker==6.1.3
filelock==3.12.4
fpdf==1.7.2
frozenlist==1.4.0
idna==3.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
lxml==4.9.3
msgpack==1.0.5
numpy==1.26.0
packaging==23.1
paramiko==3.3.1
pdfkit==1.0.0
pika==1.3.2
protobuf==4.24.3
pycparser==2.21
pymongo==4.5.0
PyNaCl==1.5.0
python-dotenv==1.0.0
python-gvm==23.5.1
PyYAML==6.0.1
ray==2.7.0
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
six==1.16.0
urllib3==2.0.5
websocket-client==1.6.3
xmltodict
zaproxy==0.1.1
